packages/core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "verify-core"
version = "0.1.0"
description = "15code Verify — core scanning engine"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [{ name = "15code", email = "dev@15code.com" }]
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.5.0",
    "tiktoken>=0.7.0",
    "tenacity>=8.2.0",
    "orjson>=3.9.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
llm = ["anthropic>=0.39.0", "openai>=1.40.0"]
dev = ["pytest>=7.4", "pytest-asyncio>=0.23", "ruff>=0.5", "mypy>=1.8"]

[tool.hatch.build.targets.wheel]
packages = ["verify_core"]
